Add userHome.js and modify HomeView.vue
