+++ frontend/src/composables/useHome.js
@@ -1,0 +1,22 @@
+import { ref } from 'vue'
+import axios from 'axios'
+
+export function useHome() {
+  const API_BASE_URL = import.meta.env.VITE_API_BASE_URL
+  const isAdmin = ref(false)
+  const containerSize = ref('')
+
+  const handleLogin = async () => {
+    const token = localStorage.getItem('token')
+    const response = await axios.get(`${API_BASE_URL}/logged_in`, {
+      headers: {
+        Authorization: `Bearer ${token}`
+      }
+    })
+
+    isAdmin.value = response.data.payload.user_id === 49
+    containerSize.value = isAdmin.value ? 'container w-50' : 'container w-75'
+  }
+
+  return { isAdmin, containerSize, handleLogin }
+}
